# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BIT College Event Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  admin: 'admin',
  faculty: 'faculty',
  student: 'student',
} as const;

export type RegisterInputDepartment = typeof RegisterInputDepartment[keyof typeof RegisterInputDepartment];


export const RegisterInputDepartment = {
  'CSE-ICB': 'CSE-ICB',
  CSE: 'CSE',
  ISE: 'ISE',
  ECE: 'ECE',
  'AI&ML': 'AI&ML',
  ME: 'ME',
  CV: 'CV',
} as const;

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
  department: RegisterInputDepartment;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  faculty: 'faculty',
  student: 'student',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  department: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type EventInputDepartment = typeof EventInputDepartment[keyof typeof EventInputDepartment];


export const EventInputDepartment = {
  'CSE-ICB': 'CSE-ICB',
  CSE: 'CSE',
  ISE: 'ISE',
  ECE: 'ECE',
  'AI&ML': 'AI&ML',
  ME: 'ME',
  CV: 'CV',
} as const;

export interface EventInput {
  /** @minLength 1 */
  title: string;
  description: string;
  date: string;
  time: string;
  venue: string;
  department: EventInputDepartment;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  rules?: string | null;
  registrationFee?: number;
  totalSlots: number;
  /** @nullable */
  paymentQrUrl?: string | null;
}

export interface EventUpdate {
  title?: string;
  description?: string;
  date?: string;
  time?: string;
  venue?: string;
  department?: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  rules?: string | null;
  registrationFee?: number;
  totalSlots?: number;
  /** @nullable */
  paymentQrUrl?: string | null;
}

export type EventApprovalStatus = typeof EventApprovalStatus[keyof typeof EventApprovalStatus];


export const EventApprovalStatus = {
  approved: 'approved',
  pending: 'pending',
  rejected: 'rejected',
} as const;

export interface EventApproval {
  status: EventApprovalStatus;
}

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


export const EventStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  time: string;
  venue: string;
  department: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  rules?: string | null;
  registrationFee: number;
  totalSlots: number;
  availableSlots: number;
  /** @nullable */
  paymentQrUrl?: string | null;
  status: EventStatus;
  /** @nullable */
  createdByFacultyId?: number | null;
  /** @nullable */
  createdByFacultyName?: string | null;
  createdAt: string;
}

export interface DepartmentStat {
  department: string;
  count: number;
}

export interface EventStats {
  totalEvents: number;
  approvedEvents: number;
  pendingEvents: number;
  totalRegistrations: number;
  totalRevenue: number;
  departmentBreakdown: DepartmentStat[];
}

export interface RegistrationInput {
  eventId: number;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
}

export type RegistrationReviewPaymentStatus = typeof RegistrationReviewPaymentStatus[keyof typeof RegistrationReviewPaymentStatus];


export const RegistrationReviewPaymentStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface RegistrationReview {
  paymentStatus: RegistrationReviewPaymentStatus;
  /** @nullable */
  adminComment?: string | null;
}

export type RegistrationPaymentStatus = typeof RegistrationPaymentStatus[keyof typeof RegistrationPaymentStatus];


export const RegistrationPaymentStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Registration {
  id: number;
  eventId: number;
  studentId: number;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  paymentStatus: RegistrationPaymentStatus;
  /** @nullable */
  adminComment?: string | null;
  registeredAt: string;
  event?: Event;
  student?: User;
}

export interface FeedbackInput {
  eventId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comments?: string | null;
  /** @nullable */
  customResponses?: string | null;
}

export interface Feedback {
  id: number;
  eventId: number;
  studentId: number;
  rating: number;
  /** @nullable */
  comments?: string | null;
  /** @nullable */
  customResponses?: string | null;
  submittedAt: string;
  student?: User;
}

export interface UploadResponse {
  url: string;
}

export interface EventReport {
  event: Event;
  totalRegistrations: number;
  approvedRegistrations: number;
  totalRevenue: number;
  registrations: Registration[];
}

export type ListEventsParams = {
department?: string;
status?: string;
search?: string;
};