# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BIT College Event Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── Events ─────────────────────────────────────────────────────────────────
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (faculty/admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/stats:
    get:
      operationId: getEventStats
      tags: [events]
      summary: Get dashboard stats for faculty/admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Event statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventStats"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (faculty/admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (faculty/admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/approve:
    patch:
      operationId: approveEvent
      tags: [events]
      summary: Approve or reject an event (admin/faculty)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventApproval"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/registrations:
    get:
      operationId: listEventRegistrations
      tags: [events]
      summary: List registrations for an event (faculty/admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
  /events/{id}/feedback:
    get:
      operationId: listEventFeedback
      tags: [events]
      summary: List feedback for an event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Feedback list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feedback"

  # ─── Registrations ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role, department]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, faculty, student]
        department:
          type: string
          enum: [CSE-ICB, CSE, ISE, ECE, AI&ML, ME, CV]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, department, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, faculty, student]
        department:
          type: string
        createdAt:
          type: string

    # ── Events ────────────────────────────────────────────────────────────────
    EventInput:
      type: object
      required: [title, description, date, time, venue, department, totalSlots]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        date:
          type: string
        time:
          type: string
        venue:
          type: string
        department:
          type: string
          enum: [CSE-ICB, CSE, ISE, ECE, AI&ML, ME, CV]
        imageUrl:
          type: ["string", "null"]
        rules:
          type: ["string", "null"]
        registrationFee:
          type: integer
          default: 0
        totalSlots:
          type: integer
        paymentQrUrl:
          type: ["string", "null"]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        time:
          type: string
        venue:
          type: string
        department:
          type: string
        imageUrl:
          type: ["string", "null"]
        rules:
          type: ["string", "null"]
        registrationFee:
          type: integer
        totalSlots:
          type: integer
        paymentQrUrl:
          type: ["string", "null"]
    EventApproval:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, pending, rejected]
    Event:
      type: object
      required: [id, title, description, date, time, venue, department, registrationFee, totalSlots, availableSlots, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
        time:
          type: string
        venue:
          type: string
        department:
          type: string
        imageUrl:
          type: ["string", "null"]
        rules:
          type: ["string", "null"]
        registrationFee:
          type: integer
        totalSlots:
          type: integer
        availableSlots:
          type: integer
        paymentQrUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdByFacultyId:
          type: ["integer", "null"]
        createdByFacultyName:
          type: ["string", "null"]
        createdAt:
          type: string
    EventStats:
      type: object
      required: [totalEvents, approvedEvents, pendingEvents, totalRegistrations, totalRevenue, departmentBreakdown]
      properties:
        totalEvents:
          type: integer
        approvedEvents:
          type: integer
        pendingEvents:
          type: integer
        totalRegistrations:
          type: integer
        totalRevenue:
          type: integer
        departmentBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentStat"
    DepartmentStat:
      type: object
      required: [department, count]
      properties:
        department:
          type: string
        count:
          type: integer

    # ── Registrations ─────────────────────────────────────────────────────────
...
    Registration:
      type: object
      required: [id, eventId, studentId, paymentStatus, registeredAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        studentId:
          type: integer
        paymentScreenshotUrl:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, approved, rejected]
        adminComment:
          type: ["string", "null"]
        registeredAt:
          type: string
        event:
          $ref: "#/components/schemas/Event"
        student:
          $ref: "#/components/schemas/User"

    # ── Feedback ──────────────────────────────────────────────────────────────
...
    Feedback:
      type: object
      required: [id, eventId, studentId, rating, submittedAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        studentId:
          type: integer
        rating:
          type: integer
        comments:
          type: ["string", "null"]
        customResponses:
          type: ["string", "null"]
        submittedAt:
          type: string
        student:
          $ref: "#/components/schemas/User"
...