# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BIT College Event Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Feedback {
  id: number;
  eventId: number;
  studentId: number;
  rating: number;
  /** @nullable */
  comments?: string | null;
  /** @nullable */
  customResponses?: string | null;
  submittedAt: string;
  student?: User;
}