# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BIT College Event Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputDepartment } from './registerInputDepartment';
import type { RegisterInputRole } from './registerInputRole';

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
  department: RegisterInputDepartment;
}