# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BIT College Event Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from './event';
import type { RegistrationPaymentStatus } from './registrationPaymentStatus';
import type { User } from './user';

export interface Registration {
  id: number;
  eventId: number;
  studentId: number;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  paymentStatus: RegistrationPaymentStatus;
  /** @nullable */
  adminComment?: string | null;
  registeredAt: string;
  event?: Event;
  student?: User;
}