# lib/db/src/schema/registrations.ts
import { pgTable, text, serial, timestamp, integer } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const registrationsTable = pgTable("registrations", {
  id: serial("id").primaryKey(),
  eventId: integer("event_id").notNull(),
  studentId: integer("student_id").notNull(),
  paymentScreenshotUrl: text("payment_screenshot_url"),
  paymentStatus: text("payment_status", { enum: ["pending", "approved", "rejected"] }).notNull().default("pending"),
  adminComment: text("admin_comment"),
  registeredAt: timestamp("registered_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertRegistrationSchema = createInsertSchema(registrationsTable).omit({ id: true, registeredAt: true });
export type InsertRegistration = z.infer<typeof insertRegistrationSchema>;
export type Registration = typeof registrationsTable.$inferSelect;